Added consent config
