+++ dpc-aggregation/src/main/resources/application.yml
@@ -91,6 +91,15 @@ bbclient:
 
   useBfdMock: ${USE_BFD_MOCK:-false}
 
+consentClient:
+  consentServiceUrl: ${CONSENT_SERVICE_URL:-"http://consent:3600/v1/"}
+  timeouts:
+    connectionTimeout: 30000 # ms
+    socketTimeout: 30000 # ms
+    requestTimeout: 30000 # ms
+
+consentHealthCheckURL: ${CONSENT_HEALTHCHECK_URL:-"http://consent:9904/healthcheck}
+
 awsQueue:
   emitAwsMetrics: ${EMIT_AWS_METRICS:-true}
   queueSizeMetricName: ${QUEUE_SIZE_METRIC_NAME:-"JobQueueBatchCount"}
@@ -100,9 +109,6 @@ awsQueue:
   awsAgeReportingInterval: ${AWS_AGE_REPORTING_INTERVAL:-600}
   environment: ${ENV:-"unknown_env"}
   awsNamespace: ${NAMESPACE:-"DPC"}
-
-consentServiceUrl: ${CONSENT_SERVICE_URL:-"http://consent:3600/v1/"}
-consentHealthCheckURL: ${CONSENT_HEALTHCHECK_URL:-"http://consent:9904/healthcheck}
 
 retryCount: 3 # Number of times to retry reach BB request
 resourcesPerFileCount: 5000 # Max number of resources that an export file will have before creating a new file
