+++ dpc-bluebutton/src/main/java/gov/cms/dpc/bluebutton/BlueButtonClientModule.java
@@ -13,6 +13,7 @@ import gov.cms.dpc.bluebutton.config.BBClientConfiguration;
 import gov.cms.dpc.bluebutton.config.BlueButtonBundleConfiguration;
 import gov.cms.dpc.bluebutton.exceptions.BlueButtonClientSetupException;
 import gov.cms.dpc.bluebutton.health.BlueButtonHealthCheck;
+import gov.cms.dpc.common.TimeoutConfiguration;
 import io.dropwizard.core.Configuration;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.config.RequestConfig;
@@ -156,7 +157,7 @@ public class BlueButtonClientModule<T extends Configuration & BlueButtonBundleCo
         }
 
         // Configure the socket timeout for the connection, incl. ssl tunneling
-        final BBClientConfiguration.TimeoutConfiguration timeouts = this.bbClientConfiguration.getTimeouts();
+        final TimeoutConfiguration timeouts = this.bbClientConfiguration.getTimeouts();
         RequestConfig requestConfig = RequestConfig.custom()
                 .setConnectTimeout(timeouts.getConnectionTimeout())
                 .setConnectionRequestTimeout(timeouts.getRequestTimeout())
